Abstract files copy functions to utils file
